Add fuzzy random number function
I don't fully understand how 'random_normal' works, though I like the
curves it creates. I don't know how to modify the results, to skew them
high or low, but this simple function will allow skewing results.

It's also easier to understand, without needing a degree in statistics.

A 'skew' of:
 1 = always add
-1 = always sub
 2 = evenly add/sub
 3 = add 2/3 of the time
-3 = sub 2/3 of the time

I'm seeing how 1 and -1 could be used; but need to also make the
'weight' or odds of when the add/sub happens a variable as well; need to
come up with a name for that? (probably one already exists)