Modify how 'icy' periods are set
Use the percent to determine if Icy Hot is enabled during each tempo
change. This is then read in by master-beater, to determine if Icy Hot
should be enabled for icy-build, if something has triggered it, by
setting /tmp/icy-add-pics to a non-zero value.

Man, shared memory is going to make a lot of this easier...