Make too late timer more sadistic
The eventual goal is to have this be a variable figure based off of the
past performances, and how long it took from being told to edge to
getting to the edge for those sessions. Something like within 1 or 2
standard deviations should be 'safe', but beyond should add edges,
etc...

roughly, if 20s, 20s, 20s, then 30s, the 30s should get penalized, or
perhaps 35s? Definitely should be a tunable, how severe it is... :D