Add voice clips words
So basically, these are used to generate voices using TTS systems and
are suggestions or prompts for various stages of the stroking game(s).

The goal is to have some occasional random clips as well of female
voices that say stuff like 'faster', 'slower', 'you call that a cock?'
...

The TTS services i've been using are:
https://www.naturalreaders.com/online/ (5 minute daily limit)
https://cloud.google.com/text-to-speech (seemingly no limit)

Then i use a Chrome extension "Chrome Audio Capture" to capture the
voice, and then i use Audacity to cut the clips out, and put them in
directories to be selected randomly when a clip matching that type is
needed.

See play_rand_sound() in edge-for-me