Game concept
Idea is a set number of images from a source directory, and during the
image-by-image slideshow, I decide to keep the image, and paddle my
balls X times, where X is the sum of points for keywords the image
matches.

For example, if "nipple" is worth 2 slaps, and there are 4 (female)
nipples in the image, then to keep the image i would need to paddle my
balls 8 times. With other keywords, like... oh I don't know, 'cock',
'cum', 'underboob', 'flashing' also match, then it becomes a
punishment...

I was playing this also with the option of avoiding the paddling, if I
could get to the edge in 30 seconds. However, if I didn't get to the
edge in 30 seconds, I still had to get to the edge, and I had to paddle
my balls, and delete the image.